+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.7.0] - 2022-04-26
 ### Added
 - Expose external API to check if the user exists. [#128](https://github.com/rokwire/health-building-block/issues/128)
 - Expose get time client API. [#130](https://github.com/rokwire/health-building-block/issues/130)
@@ -228,7 +230,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update access rule admin api.
 - Delete access rule admin api.
 
-[Unreleased]: https://github.com/rokwire/health-building-block/compare/v2.6.0...HEAD
+[Unreleased]: https://github.com/rokwire/health-building-block/compare/v2.7.0...HEAD
+[2.7.0]: https://github.com/rokwire/health-building-block/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/rokwire/health-building-block/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/rokwire/health-building-block/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/rokwire/health-building-block/compare/v2.3.0...v2.4.0